Estandarización del feature "Notification-on-login"
