+++ app/templates/sendmailbtn.html
@@ -9,7 +9,7 @@
 <body>
   <h1>Enviar Correo de Prueba</h1>
 
-  <form action="{{ url_for('email.enviar') }}" method="POST">
+  <form action="{{ url_for('mail_bp.send') }}" method="POST">
     <input type="email" name="email" placeholder="Tu correo" required>
     <button type="submit">Enviar Correo</button>
   </form>
